requirements.txt
Django==3.2.5
requests==2.22.0
djangorestframework==3.12.4
mysqlclient==2.0.3